Seguridad implmentada y falla validateToken

+++ APIMARKET/Backend/Apimarket/DTOs/HoneyCollection.DTO.cs
@@ -4,7 +4,7 @@
     {
         public int Id_HoneyCollection { get; set; }
 
-        public int Tot_HoneyCollection { get; set; }
+        public decimal Tot_HoneyCollection { get; set; }
 
         public int CanFra125_HoneyCollection { get; set; }
